Remove CDI and references to JPA

+++ tck/src/main/java/ee/jakarta/tck/data/framework/read/only/NaturalNumbers.java
@@ -29,7 +29,7 @@ import ee.jakarta.tck.data.framework.read.only.NaturalNumber.NumberType;
  * TODO figure out a way to make this a ReadOnlyRepository instead.
  */
 @Repository
-public interface NaturalNumbers extends CrudRepository<NaturalNumber, Long> {
+public interface NaturalNumbers extends CrudRepository<NaturalNumber, Integer> {
 
     Slice<NaturalNumber> findByNumTypeAndFloorOfSquareRootLessThanEqual(NumberType type,
                                                                         long maxSqrtFloor,
